reindex thread
