+++ src/components/Indexer/crawlerThread.ts
@@ -1,10 +1,23 @@
 import { parentPort, workerData } from 'worker_threads'
-import { getDeployedContractBlock, getNetworkHeight, processBlocks } from './utils.js'
+import {
+  getDeployedContractBlock,
+  getNetworkHeight,
+  processBlocks,
+  processChunkLogs
+} from './utils.js'
 import { Blockchain } from '../../utils/blockchain.js'
 import { BlocksEvents, SupportedNetwork } from '../../@types/blockchain.js'
 import { LOG_LEVELS_STR } from '../../utils/logging/Logger.js'
 import { sleep } from '../../utils/util.js'
 import { INDEXER_LOGGER } from './index.js'
+
+export interface ReindexTask {
+  txId: string
+  chainId: string
+  eventIndex?: number
+}
+
+const REINDEX_QUEUE: ReindexTask[] = []
 
 interface ThreadData {
   rpcDetails: SupportedNetwork
@@ -67,7 +80,37 @@ export async function proccesNetworkData(): Promise<void> {
         )
       }
     }
+    processReindex()
     await sleep(30000)
+  }
+}
+
+async function processReindex(): Promise<void> {
+  while (REINDEX_QUEUE.length > 0) {
+    const reindexTask = REINDEX_QUEUE.pop()
+    try {
+      const provider = blockchain.getProvider()
+      const receipt = await provider.getTransactionReceipt(reindexTask.txId)
+      if (receipt) {
+        const log = receipt.logs[reindexTask.eventIndex]
+        const logs = log ? [log] : receipt.logs
+        const events = await processChunkLogs(logs, provider, rpcDetails.chainId)
+        const eventKeys = Object.keys(events)
+        eventKeys.forEach((eventType) => {
+          INDEXER_LOGGER.logMessage(
+            `Network: ${rpcDetails.network} storing event type  ${eventType} `,
+            true
+          )
+          parentPort.postMessage({
+            method: eventType,
+            network: rpcDetails.chainId,
+            data: events[eventType]
+          })
+        })
+      }
+    } catch (error) {
+      INDEXER_LOGGER.log(LOG_LEVELS_STR.LEVEL_ERROR, `Error: ${error.message} `, true)
+    }
   }
 }
 
@@ -88,5 +131,10 @@ export async function storeFoundEvents(events: BlocksEvents): Promise<void> {
 parentPort.on('message', (message) => {
   if (message.method === 'start-crawling') {
     proccesNetworkData()
+  }
+  if (message.method === 'add-reindex-task') {
+    if (message.reindexTask) {
+      REINDEX_QUEUE.push(message.reindexTask)
+    }
   }
 })
