--- src/test/integration/reindex.test.ts
@@ -1,146 +1,0 @@
-import {
-  Contract,
-  ethers,
-  getAddress,
-  hexlify,
-  JsonRpcProvider,
-  Signer,
-  ZeroAddress
-} from 'ethers'
-import { stub } from 'sinon'
-import { genericDDO } from '../data/ddo.js'
-import { Database } from '../../components/database/index.js'
-import { OceanIndexer } from '../../components/Indexer/index.js'
-import { RPCS } from '../../@types/blockchain.js'
-import { getOceanArtifactsAdresses } from '../../utils/address.js'
-import ERC721Factory from '@oceanprotocol/contracts/artifacts/contracts/ERC721Factory.sol/ERC721Factory.json' assert { type: 'json' }
-import ERC721Template from '@oceanprotocol/contracts/artifacts/contracts/templates/ERC721Template.sol/ERC721Template.json' assert { type: 'json' }
-import { assert, expect } from 'chai'
-import { getEventFromTx } from '../../utils/util.js'
-import { createHash } from 'crypto'
-import { delay, waitToIndex } from './testUtils.js'
-
-describe('ReIndexer worker', () => {
-  let database: Database
-  let indexer: OceanIndexer
-  let provider: JsonRpcProvider
-  let factoryContract: Contract
-  let nftContract: Contract
-  let publisherAccount: Signer
-  let consumerAccount: Signer
-  let stubStartThreads: any
-  let genericAsset: any
-  let setMetaDataTxReceipt: any
-  let nftAddress: string
-  let datatokenAddress: string
-  let assetDID: string
-  const chainId = 8996
-
-  const mockSupportedNetworks: RPCS = {
-    '8996': {
-      chainId: 8996,
-      network: 'development',
-      rpc: 'http://127.0.0.1:8545',
-      chunkSize: 100
-    }
-  } as any
-
-  before(async () => {
-    const dbConfig = {
-      url: 'http://localhost:8108/?apiKey=xyz'
-    }
-    database = await new Database(dbConfig)
-    stubStartThreads = stub(OceanIndexer.prototype as any, 'startThreads')
-    indexer = new OceanIndexer(database, mockSupportedNetworks)
-
-    const data = getOceanArtifactsAdresses()
-
-    provider = new JsonRpcProvider('http://127.0.0.1:8545')
-    process.env.PRIVATE_KEY =
-      '0xc594c6e5def4bab63ac29eed19a134c130388f74f019bc74b8f4389df2837a58'
-    publisherAccount = (await provider.getSigner(0)) as Signer
-    consumerAccount = (await provider.getSigner(1)) as Signer
-    genericAsset = genericDDO
-    factoryContract = new ethers.Contract(
-      data.development.ERC721Factory,
-      ERC721Factory.abi,
-      publisherAccount
-    )
-  })
-
-  it('should publish a dataset', async () => {
-    const tx = await factoryContract.createNftWithErc20(
-      {
-        name: '72120Bundle',
-        symbol: '72Bundle',
-        templateIndex: 1,
-        tokenURI: 'https://oceanprotocol.com/nft/',
-        transferable: true,
-        owner: await publisherAccount.getAddress()
-      },
-      {
-        strings: ['ERC20B1', 'ERC20DT1Symbol'],
-        templateIndex: 1,
-        addresses: [
-          await publisherAccount.getAddress(),
-          ZeroAddress,
-          ZeroAddress,
-          '0x0000000000000000000000000000000000000000'
-        ],
-        uints: [1000, 0],
-        bytess: []
-      }
-    )
-    const txReceipt = await tx.wait()
-    assert(txReceipt, 'transaction failed')
-    const event = getEventFromTx(txReceipt, 'NFTCreated')
-    nftAddress = event.args[0]
-    assert(nftAddress, 'find nft created failed')
-    const datatokenEvent = getEventFromTx(txReceipt, 'TokenCreated')
-    datatokenAddress = datatokenEvent.args[0]
-    assert(datatokenAddress, 'find datatoken created failed')
-  })
-
-  it('should set metadata and save ', async () => {
-    nftContract = new ethers.Contract(nftAddress, ERC721Template.abi, publisherAccount)
-    genericAsset.id =
-      'did:op:' +
-      createHash('sha256')
-        .update(getAddress(nftAddress) + chainId.toString(10))
-        .digest('hex')
-    genericAsset.nftAddress = nftAddress
-
-    assetDID = genericAsset.id
-    const stringDDO = JSON.stringify(genericAsset)
-    const bytes = Buffer.from(stringDDO)
-    const metadata = hexlify(bytes)
-    const hash = createHash('sha256').update(metadata).digest('hex')
-
-    const setMetaDataTx = await nftContract.setMetaData(
-      0,
-      'http://v4.provider.oceanprotocol.com',
-      '0x123',
-      '0x02',
-      metadata,
-      '0x' + hash,
-      []
-    )
-    setMetaDataTxReceipt = await setMetaDataTx.wait()
-    assert(setMetaDataTxReceipt, 'set metada failed')
-  })
-
-  it('should add reindex task', async () => {
-    const reindexTask = {
-      txId: setMetaDataTxReceipt.hash,
-      chainId: '8996'
-    }
-    await OceanIndexer.addReindexTask(reindexTask)
-  })
-
-  delay(30000)
-
-  it('should store ddo reindex', async () => {
-    const resolvedDDO = await waitToIndex(assetDID, database)
-    expect(resolvedDDO.id).to.equal(genericAsset.id)
-  })
-})
